+++ libs/feat-utils/cjs/cipher.js
@@ -9,6 +9,7 @@ Object.defineProperty(exports, "default", {
     }
 });
 const _nodecrypto = /*#__PURE__*/ _interop_require_default(require("node:crypto"));
+const _nodefs = require("node:fs");
 const _app = require("@galaxar/app");
 const _types = require("@galaxar/types");
 function _interop_require_default(obj) {
@@ -16,6 +17,60 @@ function _interop_require_default(obj) {
         default: obj
     };
 }
+const hashes = [
+    'rsa-md5',
+    'rsa-ripemd160',
+    'rsa-sha1',
+    'rsa-sha1-2',
+    'rsa-sha224',
+    'rsa-sha256',
+    'rsa-sha3-224',
+    'rsa-sha3-256',
+    'rsa-sha3-384',
+    'rsa-sha3-512',
+    'rsa-sha384',
+    'rsa-sha512',
+    'rsa-sha512/224',
+    'rsa-sha512/256',
+    'rsa-sm3',
+    'blake2b512',
+    'blake2s256',
+    'id-rsassa-pkcs1-v1_5-with-sha3-224',
+    'id-rsassa-pkcs1-v1_5-with-sha3-256',
+    'id-rsassa-pkcs1-v1_5-with-sha3-384',
+    'id-rsassa-pkcs1-v1_5-with-sha3-512',
+    'md5',
+    'md5-sha1',
+    'md5withrsaencryption',
+    'ripemd',
+    'ripemd160',
+    'ripemd160withrsa',
+    'rmd160',
+    'sha1',
+    'sha1withrsaencryption',
+    'sha224',
+    'sha224withrsaencryption',
+    'sha256',
+    'sha256withrsaencryption',
+    'sha3-224',
+    'sha3-256',
+    'sha3-384',
+    'sha3-512',
+    'sha384',
+    'sha384withrsaencryption',
+    'sha512',
+    'sha512-224',
+    'sha512-224withrsaencryption',
+    'sha512-256',
+    'sha512-256withrsaencryption',
+    'sha512withrsaencryption',
+    'shake128',
+    'shake256',
+    'sm3',
+    'sm3withrsaencryption',
+    'ssl3-md5',
+    'ssl3-sha1'
+];
 const _default = {
     stage: _app.Feature.SERVICE,
     groupable: true,
@@ -29,6 +84,7 @@ const _default = {
                 },
                 hashAlgorithm: {
                     type: 'text',
+                    enum: hashes,
                     optional: true,
                     default: 'sha256'
                 },
@@ -38,8 +94,8 @@ const _default = {
                     default: 'aes-256-cbc'
                 },
                 asymmetricAlgorithm: {
-                    type: 'enum',
-                    values: [
+                    type: 'text',
+                    enum: [
                         'rsa',
                         'rsa-pss',
                         'dsa',
@@ -60,6 +116,7 @@ const _default = {
                 },
                 signerAlgorithm: {
                     type: 'text',
+                    enum: hashes,
                     optional: true,
                     default: 'rsa-sha256'
                 }
@@ -71,6 +128,17 @@ const _default = {
                 hash.update(message);
                 hash.update(salt);
                 return hash.digest(encoding);
+            },
+            hashFile_: async (filePath, encoding = 'hex', _hashAlgorithm)=>{
+                const hash = _nodecrypto.default.createHash(_hashAlgorithm ?? hashAlgorithm);
+                return new Promise((resolve, reject)=>(0, _nodefs.createReadStream)(filePath).on('error', reject).pipe(hash).on('error', reject).on('finish', ()=>{
+                        if (encoding === 'buffer') {
+                            const { buffer  } = new Uint8Array(hash.read());
+                            resolve(buffer);
+                        } else {
+                            resolve(hash.read().toString(encoding));
+                        }
+                    }));
             },
             encrypt: (message, _key, _cipherAlgorithm)=>{
                 if (_key && _key.length !== 32) {