+++ libs/feat-utils/cjs/index.js
@@ -12,20 +12,28 @@ _export(exports, {
     cipher: function() {
         return _cipher.default;
     },
-    hasher: function() {
-        return _hasher.default;
-    },
     imageProcessor: function() {
         return _imageProcessor.default;
     },
+    threadPool: function() {
+        return _threadPool.default;
+    },
+    threadWorker: function() {
+        return _threadWorker.default;
+    },
     ttlMemCache: function() {
         return _ttlMemCache.default;
+    },
+    webSocketClient: function() {
+        return _webSocketClient.default;
     }
 });
 const _cipher = /*#__PURE__*/ _interop_require_default(require("./cipher"));
-const _hasher = /*#__PURE__*/ _interop_require_default(require("./hasher"));
 const _imageProcessor = /*#__PURE__*/ _interop_require_default(require("./imageProcessor"));
+const _threadPool = /*#__PURE__*/ _interop_require_default(require("./threadPool"));
+const _threadWorker = /*#__PURE__*/ _interop_require_default(require("./threadWorker"));
 const _ttlMemCache = /*#__PURE__*/ _interop_require_default(require("./ttlMemCache"));
+const _webSocketClient = /*#__PURE__*/ _interop_require_default(require("./webSocketClient"));
 function _interop_require_default(obj) {
     return obj && obj.__esModule ? obj : {
         default: obj