+++ libs/feat-utils/lib/cipher.js
@@ -1,6 +1,62 @@
 import crypto from 'node:crypto';
+import { createReadStream } from 'node:fs';
 import { Feature } from '@galaxar/app';
 import { ValidationError } from '@galaxar/types';
+
+const hashes = [
+    'rsa-md5',
+    'rsa-ripemd160',
+    'rsa-sha1',
+    'rsa-sha1-2',
+    'rsa-sha224',
+    'rsa-sha256',
+    'rsa-sha3-224',
+    'rsa-sha3-256',
+    'rsa-sha3-384',
+    'rsa-sha3-512',
+    'rsa-sha384',
+    'rsa-sha512',
+    'rsa-sha512/224',
+    'rsa-sha512/256',
+    'rsa-sm3',
+    'blake2b512',
+    'blake2s256',
+    'id-rsassa-pkcs1-v1_5-with-sha3-224',
+    'id-rsassa-pkcs1-v1_5-with-sha3-256',
+    'id-rsassa-pkcs1-v1_5-with-sha3-384',
+    'id-rsassa-pkcs1-v1_5-with-sha3-512',
+    'md5',
+    'md5-sha1',
+    'md5withrsaencryption',
+    'ripemd',
+    'ripemd160',
+    'ripemd160withrsa',
+    'rmd160',
+    'sha1',
+    'sha1withrsaencryption',
+    'sha224',
+    'sha224withrsaencryption',
+    'sha256',
+    'sha256withrsaencryption',
+    'sha3-224',
+    'sha3-256',
+    'sha3-384',
+    'sha3-512',
+    'sha384',
+    'sha384withrsaencryption',
+    'sha512',
+    'sha512-224',
+    'sha512-224withrsaencryption',
+    'sha512-256',
+    'sha512-256withrsaencryption',
+    'sha512withrsaencryption',
+    'shake128',
+    'shake256',
+    'sm3',
+    'sm3withrsaencryption',
+    'ssl3-md5',
+    'ssl3-sha1',
+];
 
 export default {
     stage: Feature.SERVICE,
@@ -14,16 +70,16 @@ export default {
                 {
                     schema: {
                         key: { type: 'text', fixedLength: 32, optional: true },
-                        hashAlgorithm: { type: 'text', optional: true, default: 'sha256' },
+                        hashAlgorithm: { type: 'text', enum: hashes, optional: true, default: 'sha256' },
                         cipherAlgorithm: { type: 'text', optional: true, default: 'aes-256-cbc' },
                         asymmetricAlgorithm: {
-                            type: 'enum',
-                            values: ['rsa', 'rsa-pss', 'dsa', 'ec', 'ed25519', 'ed448', 'x25519', 'x448', 'dh'],
+                            type: 'text',
+                            enum: ['rsa', 'rsa-pss', 'dsa', 'ec', 'ed25519', 'ed448', 'x25519', 'x448', 'dh'],
                             optional: true,
                             default: 'rsa',
                         },
                         asymmetricBits: { type: 'integer', optional: true, default: 2048 },
-                        signerAlgorithm: { type: 'text', optional: true, default: 'rsa-sha256' },
+                        signerAlgorithm: { type: 'text', enum: hashes, optional: true, default: 'rsa-sha256' },
                     },
                 },
                 name
@@ -36,6 +92,24 @@ export default {
                 hash.update(salt);
 
                 return hash.digest(encoding);
+            },
+
+            hashFile_: async (filePath, encoding = 'hex', _hashAlgorithm) => {
+                const hash = crypto.createHash(_hashAlgorithm ?? hashAlgorithm);
+                return new Promise((resolve, reject) =>
+                    createReadStream(filePath)
+                        .on('error', reject)
+                        .pipe(hash)
+                        .on('error', reject)
+                        .on('finish', () => {
+                            if (encoding === 'buffer') {
+                                const { buffer } = new Uint8Array(hash.read());
+                                resolve(buffer);
+                            } else {
+                                resolve(hash.read().toString(encoding));
+                            }
+                        })
+                );
             },
 
             encrypt: (message, _key, _cipherAlgorithm) => {
@@ -141,7 +215,7 @@ export default {
                 verifier.update(message);
 
                 return verifier.verify(publicKey, signature, encoding);
-            }
+            },
         };
 
         app.registerService(name, service);