+++ libs/feat-utils/lib/index.js
@@ -1,4 +1,6 @@
 export { default as cipher } from './cipher';
-export { default as hasher } from './hasher';
 export { default as imageProcessor } from './imageProcessor';
+export { default as threadPool } from './threadPool';
+export { default as threadWorker } from './threadWorker';
 export { default as ttlMemCache } from './ttlMemCache';
+export { default as webSocketClient } from './webSocketClient';